moved the airtable code to a different file
